Move `ShapeContext` and segment shaping into `parley_core`
Font selection remains in `parley`.
