--- parley/src/convert.rs
@@ -1,7 +1,0 @@
-// Copyright 2024 the Parley Authors
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-pub(crate) fn script_to_harfrust(script: fontique::Script) -> harfrust::Script {
-    harfrust::Script::from_iso15924_tag(harfrust::Tag::new(&script.to_bytes()))
-        .unwrap_or(harfrust::script::UNKNOWN)
-}
