+++ parley/src/lib.rs
@@ -111,10 +111,8 @@ pub use fontique;
 mod analysis;
 mod builder;
 mod context;
-mod convert;
 mod font;
 mod inline_box;
-mod lru_cache;
 mod resolve;
 mod shape;
 mod util;
