+++ parley_core/src/lib.rs
@@ -29,8 +29,10 @@ pub mod bidi;
 pub mod break_overrides;
 mod glyph;
 pub mod itemize;
+mod lru_cache;
 pub mod shape;
 
 pub use analysis::{Analysis, AnalysisDataSources, Boundary, CharInfo};
 pub use analyzer::{AnalysisOptions, Analyzer};
 pub use glyph::Glyph;
+pub use shape::shaper::{FontInstance, ShapeContext, ShapeOptions};
